Updated notebook to reflect Team Meeting Discussion on 1/2/2022
This is the notebook and python code used to generate our new ML dataset. In our team meeting we brainstormed some ways we could improve our ML model. One of my suggestions was to add back in some independent variables we previously filtered out. Additionally, after some further data exploration in Tableau, we decided to narrow the scope of the data to the top 5 crime type categories. The attached csv now contains 508,364 rows of data and encompasses the years 2017-2020, where the crime type falls into one of the following 5 categories and contains no NaN values:
•	Destruction/Damage/Vandalism of Property
•	Burglary/Breaking & Entering
•	Assault Offenses
•	Drug/Narcotic Offenses
•	Larceny/Theft Offenses